close socket on server failure
when something in the main server loop fails, will now close connection
on server side. we should rethink the whole socket thing though, like
who owns / is responsible for a socket. this is an easy fix now to get
unidirectional stuff working. things just throw errors now in that case.