Lots of changes
made long standing connections, so sockets aren't opened each request.
sockets are assumed to be unidirectional.

reorganized the network interface to hold all of the open sockets and listen
for new incoming sockets

added a messagelistener to listen to a specific incoming socket. the incoming
message can be either a request, which is handled by the noderuntime's requesthandler
(formerly serverhandler), or can be a response, in which it is sent to a responsequeue
in the client.